pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rda-service"
version = "0.1.0"
description = "Local OCR + LLM contract extraction service"
authors = [{name = "RDA", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pdfplumber>=0.11",
    "pypdf>=4.0",
    "pytesseract>=0.3",
    "Pillow>=10.0",
    "requests>=2.31",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
